+++ cli/cmd/run/run.go
@@ -30,7 +30,7 @@ import (
 	"github.com/docker/compose-cli/api/containers"
 	"github.com/docker/compose-cli/cli/options/run"
 	"github.com/docker/compose-cli/context/store"
-	"github.com/docker/compose-cli/progress"
+	"github.com/docker/compose-cli/api/progress"
 )
 
 // Command runs a container
