+++ cli/cmd/volume/command.go
@@ -26,7 +26,7 @@ import (
 	"github.com/docker/compose-cli/context/store"
 	"github.com/docker/compose-cli/ecs"
 	formatter2 "github.com/docker/compose-cli/formatter"
-	"github.com/docker/compose-cli/progress"
+	"github.com/docker/compose-cli/api/progress"
 
 	"github.com/hashicorp/go-multierror"
 	"github.com/spf13/cobra"
