+++ local/compose/down.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/docker/compose-cli/api/compose"
 
-	"github.com/docker/compose-cli/progress"
+	"github.com/docker/compose-cli/api/progress"
 
 	"github.com/compose-spec/compose-go/cli"
 	"github.com/compose-spec/compose-go/types"
